refactor(homepage): move header component into header folder to add more child components
