+++ homepage/src/layouts/pageWrapper.jsx
@@ -1,5 +1,5 @@
 import Footer from './footer/footer';
-import Header from './header';
+import Header from './header/header';
 
 const PageWrapper = ({ children, nav, siteData }) => (
   <>
